Further submodules for transformation common module: - ui functions separated to edit based and common modules - field click functions moved to ui edit module - show transformed now available via global window object - cleaned up common script (just about only inits and module export remaining)
